add list using objects
