+++ src/index.html
@@ -14,41 +14,32 @@
   </head>
   <body>
     <main>
-      <article class="grocery-wrapper">
-        <p class="alert"></p>
-        <header class="heading">
-          <h2 class="grocery-title">Today's Todo</h2>
-          <span><i class="fa-solid fa-rotate"></i></span>
-        </header>
-        <form class="grocery-form">
+      <header class="heading">
+        <h3 class="app-title">Today's ToDo</h3>
+        <span><i class="fa-solid fa-rotate"></i></span>
+      </header>
+
+      <div class="form-container">
+        <form class="Todo-form">
           <input
             type="text"
-            class="grocery-item"
-            placeholder="Add to your list"
-          />
+            placeholder="Add to your list..."
+            class="add-text"
+          /><br />
         </form>
-        <div class="grocery-container">
-          <div class="grocery-list"></div>
-          <button class="clear-list">clear items</button>
+        <div class="label-container">
+          <!-- <div class="label">
+           
+          </div> -->
         </div>
-      </article>
+        <!-- <div class="label">
+          <input type="checkbox" value="dishes" />
+          <label>complete the todo list project</label><br />
+        </div> -->
+        <p class="clear-list">clear all completed</p>
+      </div>
     </main>
 
     <script type="module" src="./index.js"></script>
   </body>
 </html>
-<!-- <!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./styles.css" />
-   
-    <title>Grocery_App</title>
-  </head>
-  <body>
-    
-    <script src="./app.js"></script>
-  </body>
-</html> -->
