+++ src/style.css
@@ -5,140 +5,54 @@
 }
 
 body {
-  background-color: #f4f4f4;
+  display: flex;
+  justify-content: center;
 }
 
 main {
-  margin: 200px 10px 0 10px;
-}
-
-.grocery-wrapper {
-  max-width: 600px;
-  margin: 0 auto;
+  margin-top: 200px;
+  width: 400px;
   background-color: #fff;
-  box-shadow: 0 0 5px #c9c6c6;
-  padding: 30px 0;
-  border-radius: 5px;
-
-  /* position: relative; */
-}
-
-.alert {
-  width: 70%;
-  margin: 0 auto;
-  border: none;
-  border-radius: 5px;
-  text-align: center;
-  padding: 2px 0;
-  color: #fff;
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-  text-transform: capitalize;
-  position: absolute;
-  top: 150px;
-}
-
-.alert-danger {
-  display: block;
-  background-color: rgba(247, 176, 176, 0.438);
-  color: rgb(114, 4, 4);
-}
-
-.alert-success {
-  display: block;
-  background-color: rgb(195, 235, 195);
-  color: rgb(3, 97, 3);
+  box-shadow: 0 0 10px #dddbdb;
 }
 
 .heading {
   display: flex;
   justify-content: space-between;
-  align-items: center;
-  border-bottom: 1px solid #777575a2;
-  padding: 0 20px;
+  padding: 10px;
+  border-bottom: 1px solid #858383be;
 }
 
-.grocery-title {
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-  font-size: 22px;
-  line-height: 42px;
+.form-container {
+  margin-top: 10px;
 }
 
-.grocery-form {
-  position: relative;
-  margin-bottom: 15px;
-  border-bottom: 1px solid #777575a2;
-}
-
-.grocery-item {
-  padding: 8px;
-  width: 80%;
-  margin: 0 auto;
-  font-size: 16px;
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+.add-text {
+  width: 400px;
+  padding: 5px 10px;
+  font-style: italic;
   border: none;
+  border-bottom: 1px solid #858383be;
+  margin-bottom: 10px;
   outline: none;
 }
 
-.grocery-container {
-  text-align: center;
-  visibility: hidden;
-
-  /* border-bottom: 1px solid #777575a2; */
+.label {
+  border-bottom: 1px solid #858383be;
+  margin-bottom: 10px;
+  padding: 5px 10px;
 }
 
-.show-container {
-  visibility: visible;
+.icon-more {
+  float: right;
 }
 
-/* .grocery-list {
+.icon-trash {
+  float: right;
   display: none;
-} */
-
-/* .show-grocery-list {
-  display: block;
-} */
-
-.listed-item {
-  display: flex;
-  justify-content: space-between;
-
-  /* margin: 0 20px; */
-  padding: 0 20px;
-  margin-bottom: 15px;
-  border-bottom: 1px solid #777575a2;
-}
-
-.selected-item {
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-  font-size: 16px;
-  line-height: 24px;
-  letter-spacing: 0.8px;
-  text-transform: capitalize;
-}
-
-.edit-btn,
-.close-btn {
-  color: #777575;
-  background-color: transparent;
-  border: none;
-}
-
-.close-btn {
-  color: #777575;
 }
 
 .clear-list {
   text-align: center;
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-  letter-spacing: 2px;
-  background-color: transparent;
-  border: none;
-}
-
-@media screen and (min-width: 768px) {
-  .alert {
-    max-width: 40%;
-    margin: 0 auto;
-    left: 415px;
-  }
+  margin: 20px 0 10px 0;
 }
